authentication

+++ backend/src/main/java/com/app/backend/services/AuthService.java
@@ -3,6 +3,7 @@ package com.app.backend.services;
 import com.app.backend.dto.*;
 import com.app.backend.entities.User;
 import com.app.backend.exceptions.BadRequestExceptionHandler;
+import com.app.backend.exceptions.UnAuthorizeExceptionHandler;
 import com.app.backend.utils.CookieUtil;
 import com.app.backend.utils.MailUtil;
 import com.app.backend.utils.TokenUtil;
@@ -19,6 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AuthService {
     private final TokenService tokenService;
     private final UserService userService;
+    private final CacheService cacheService;
     private final MailUtil mailUtil;
     private final TokenUtil tokenUtil;
     private final CookieUtil cookieUtil;
@@ -38,6 +40,7 @@ public class AuthService {
         User user=userService.create(request);
         String accessToken=tokenUtil.generateAccessToken(user);
         String refreshToken=tokenUtil.generateRefreshToken(user);
+        cacheService.saveToken(accessToken, user.getUsername());
         tokenService.create(user,refreshToken);
         cookieUtil.addCookie(refreshToken,response);
         mailUtil.sendMail(user.getEmail(), "Hi "+user.getUsername(),"Your account was successfully registered");
@@ -50,6 +53,7 @@ public class AuthService {
         userService.matchesPasswords(request.getPassword(), user.getPassword());
         String accessToken=tokenUtil.generateAccessToken(user);
         String refreshToken=tokenUtil.generateRefreshToken(user);
+        cacheService.saveToken(accessToken, user.getUsername());
         tokenService.authToken(refreshToken,user);
         cookieUtil.addCookie(refreshToken,response);
         return authResponse(user,accessToken);
@@ -60,11 +64,11 @@ public class AuthService {
         if(!tokenUtil.validateToken(refreshToken)){
             throw new BadRequestExceptionHandler("Invalid token");
         }
-        String username=tokenUtil.extractSubject(refreshToken);
-        User user=userService.findSubject(username);
+        User user=tokenService.findByRefreshToken(refreshToken).getUser();
         tokenService.deleteToken(refreshToken);
         String newAccessToken=tokenUtil.generateAccessToken(user);
         String newRefreshToken=tokenUtil.generateRefreshToken(user);
+        cacheService.saveToken(newAccessToken, user.getUsername());
         tokenService.authToken(newRefreshToken,user);
         cookieUtil.addCookie(newRefreshToken,response);
         return authResponse(user,newAccessToken);
@@ -72,8 +76,12 @@ public class AuthService {
 
     @Transactional
     public void logout(String refreshToken,HttpServletResponse response){
+        if (!tokenUtil.validateToken(refreshToken)) {
+            throw new BadRequestExceptionHandler("Invalid refresh token");
+        }
         String username=tokenUtil.extractSubject(refreshToken);
         userService.findSubject(username);
+        cacheService.removeTokenByUsername(username);
         tokenService.deleteToken(refreshToken);
         cookieUtil.clearCookie(response);
     }
@@ -88,15 +96,19 @@ public class AuthService {
     @Transactional
     public String resetPassword(ResetPasswordRequest request){
         if (!tokenUtil.validateToken(request.getToken())){
-            throw new BadRequestExceptionHandler("Invalid token");
+            throw new UnAuthorizeExceptionHandler("Invalid token");
         }
         String subject=tokenUtil.extractSubject(request.getToken());
         User user=userService.findSubject(subject);
         userService.updatePassword(request.getPassword(),user);
+        cacheService.removeToken(request.getToken());
         return "Password updated successfully";
     }
     public CurrentUser getCurrentUser(){
         Authentication authentication= SecurityContextHolder.getContext().getAuthentication();
+        if(authentication == null || !authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")){
+            throw new UnAuthorizeExceptionHandler("User not authenticated!");
+        }
         User user=(User) authentication.getPrincipal();
         return new CurrentUser(user.getId(), user.getUsername(), user.getEmail(), user.getRole());
     }
